Improved encounter- and webhook whitelist file parsing.
This allows Pokemon names or IDs to be present in whitelist files. Lines
starting with “#” or “-“ are being ignored, so basically you have a
combined black- and whitelist. Also removed separate blacklists and the
possibility to give lists directly via commandline, so -wwht, -wblk and
-wblkf are gone!
Added 2nd Part, Encounters Were Done Added Webhooks~